Using maven + Register as external jar
Maven is still not ready, couldn't find repositories for bukkit so it's
still a WIP, that's why I have uploaded it to the experimental branch.
As soon as it is ready I'll upload it to master.
The reason I'm using Register as an external jar is that I don't see
why should it be inside of every jar, it can be just another plugin,
and that's the way nijikokun wants it to be.